requirements.txt
django~=3.1.3
django_countries
django_rest_framework
django_cors_headers
pandas~=1.1.4
joblib~=0.17.0
numpy
scikit-learn~=0.24.0
djangorestframework~=3.12.2
pytest~=6.1.2